requirements.txt
functions-framework==3.8.1
Flask==3.0.3
google-cloud-storage==2.14.0